chore(dataplane): remove CommonUserService
