+++ gravitee-am-management-api/gravitee-am-management-api-service/src/main/java/io/gravitee/am/management/service/impl/OrganizationUserServiceImpl.java
@@ -18,10 +18,13 @@ package io.gravitee.am.management.service.impl;
 import com.google.common.base.Strings;
 import io.gravitee.am.business.user.UpdateUsernameOrganizationRule;
 import io.gravitee.am.common.audit.EventType;
+import io.gravitee.am.common.oidc.StandardClaims;
 import io.gravitee.am.common.utils.RandomString;
+import io.gravitee.am.identityprovider.api.DefaultUser;
+import io.gravitee.am.identityprovider.api.UserProvider;
+import io.gravitee.am.management.service.IdentityProviderManager;
 import io.gravitee.am.management.service.OrganizationUserService;
 import io.gravitee.am.model.AccountAccessToken;
-import io.gravitee.am.model.Application;
 import io.gravitee.am.model.Organization;
 import io.gravitee.am.model.Reference;
 import io.gravitee.am.model.ReferenceType;
@@ -30,31 +33,41 @@ import io.gravitee.am.model.UserId;
 import io.gravitee.am.model.common.Page;
 import io.gravitee.am.model.membership.MemberType;
 import io.gravitee.am.repository.management.api.search.FilterCriteria;
+import io.gravitee.am.service.AuditService;
+import io.gravitee.am.service.MembershipService;
+import io.gravitee.am.service.PasswordService;
 import io.gravitee.am.service.authentication.crypto.password.bcrypt.BCryptPasswordEncoder;
 import io.gravitee.am.service.exception.InvalidParameterException;
 import io.gravitee.am.service.exception.InvalidPasswordException;
 import io.gravitee.am.service.exception.InvalidUserException;
-import io.gravitee.am.service.exception.NotImplementedException;
 import io.gravitee.am.service.exception.UserAlreadyExistsException;
 import io.gravitee.am.service.exception.UserInvalidException;
 import io.gravitee.am.service.exception.UserNotFoundException;
 import io.gravitee.am.service.exception.UserProviderNotFoundException;
+import io.gravitee.am.service.impl.PasswordHistoryService;
+import io.gravitee.am.service.model.AbstractNewUser;
 import io.gravitee.am.service.model.NewAccountAccessToken;
 import io.gravitee.am.service.model.NewOrganizationUser;
+import io.gravitee.am.service.model.UpdateUser;
 import io.gravitee.am.service.reporter.builder.AuditBuilder;
 import io.gravitee.am.service.reporter.builder.management.UserAuditBuilder;
+import io.gravitee.am.service.validators.user.UserValidator;
 import io.reactivex.rxjava3.core.Completable;
 import io.reactivex.rxjava3.core.Flowable;
 import io.reactivex.rxjava3.core.Maybe;
 import io.reactivex.rxjava3.core.Single;
+import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
+import java.time.Instant;
 import java.util.Date;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.Optional;
-import java.util.function.BiFunction;
 
+import static com.google.common.base.Strings.isNullOrEmpty;
 import static io.gravitee.am.management.service.impl.IdentityProviderManagerImpl.IDP_GRAVITEE;
 import static io.gravitee.am.model.ReferenceType.ORGANIZATION;
 import static org.springframework.util.StringUtils.hasText;
@@ -64,21 +77,34 @@ import static org.springframework.util.StringUtils.hasText;
  * @author GraviteeSource Team
  */
 @Component("managementOrganizationUserService")
-public class OrganizationUserServiceImpl extends AbstractUserService<io.gravitee.am.service.OrganizationUserService> implements OrganizationUserService {
+@Slf4j
+public class OrganizationUserServiceImpl implements OrganizationUserService {
 
     public static final BCryptPasswordEncoder PWD_ENCODER = new BCryptPasswordEncoder();
 
     @Autowired
     private io.gravitee.am.service.OrganizationUserService userService;
 
-    @Override
+    @Autowired
+    protected IdentityProviderManager identityProviderManager;
+
+    @Autowired
+    protected PasswordService passwordService;
+
+    @Autowired
+    protected UserValidator userValidator;
+
+    @Autowired
+    protected AuditService auditService;
+
+    @Autowired
+    protected MembershipService membershipService;
+
+    @Autowired
+    protected PasswordHistoryService passwordHistoryService;
+
     protected io.gravitee.am.service.OrganizationUserService getUserService() {
         return this.userService;
-    }
-
-    @Override
-    protected BiFunction<String, String, Maybe<Application>> checkClientFunction() {
-        return (x, y) -> Maybe.error(new NotImplementedException());
     }
 
     @Override
@@ -113,12 +139,6 @@ public class OrganizationUserServiceImpl extends AbstractUserService<io.gravitee
                             .doOnSuccess(user1 -> auditService.report(AuditBuilder.builder(UserAuditBuilder.class).type(EventType.USER_CREATED).user(user1)))
                             .doOnError(err -> auditService.report(AuditBuilder.builder(UserAuditBuilder.class).type(EventType.USER_CREATED).reference(new Reference(referenceType, referenceId)).throwable(err)));
                 }));
-    }
-
-    protected User transform(NewOrganizationUser newUser, ReferenceType referenceType, String referenceId) {
-        var user = super.transform(newUser, referenceType, referenceId);
-        user.setServiceAccount(newUser.isServiceAccount());
-        return user;
     }
 
     public Single<User> createGraviteeUser(Organization organization, NewOrganizationUser newUser, io.gravitee.am.identityprovider.api.User principal) {
@@ -324,5 +344,188 @@ public class OrganizationUserServiceImpl extends AbstractUserService<io.gravitee
     @Override
     public Single<User> updateStatus(String organizationId, String userId, boolean status, io.gravitee.am.identityprovider.api.User principal) {
         return updateStatus(ORGANIZATION, organizationId, userId, status, principal);
+    }
+
+    @Override
+    public Single<User> findById(ReferenceType referenceType, String referenceId, String id) {
+        return getUserService().findById(referenceType, referenceId, id);
+    }
+
+    @Override
+    public Single<User> update(ReferenceType referenceType, String referenceId, String id, UpdateUser updateUser, io.gravitee.am.identityprovider.api.User principal) {
+        return this.updateUser(referenceType, referenceId, id, updateUser, principal);
+    }
+
+    private Single<User> updateUser(ReferenceType referenceType, String referenceId, String id, UpdateUser updateUser, io.gravitee.am.identityprovider.api.User principal) {
+        if (id == null) {
+            return Single.error(new InvalidUserException("User id is required"));
+        }
+        return getUserService().findById(referenceType, referenceId, id).flatMap(user -> {
+
+                    if (Boolean.FALSE.equals(user.isInternal()) && Boolean.TRUE.equals(updateUser.getForceResetPassword())) {
+                        return Single.error(new InvalidUserException("forceResetPassword is forbidden on external users"));
+                    }
+
+                    // This handles identity providers not enforcing email
+                    final boolean validateEmailIfNecessary = !(isNullOrEmpty(user.getEmail()) && isNullOrEmpty(updateUser.getEmail()));
+                    return userValidator.validate(updateUser, validateEmailIfNecessary).andThen(Single.just(user));
+                })
+                .flatMap(user -> updateWithProviderIfNecessary(updateUser, user))
+                .flatMap(user -> getUserService().update(referenceType, referenceId, id, user)
+                        .doOnSuccess(user1 -> auditService.report(AuditBuilder.builder(UserAuditBuilder.class).principal(principal).type(EventType.USER_UPDATED).oldValue(user).user(user1)))
+                        .doOnError(throwable -> auditService.report(AuditBuilder.builder(UserAuditBuilder.class).principal(principal).type(EventType.USER_UPDATED).reference(new Reference(referenceType, referenceId)).throwable(throwable))));
+    }
+
+    private Single<UpdateUser> updateWithProviderIfNecessary(UpdateUser updateUser, User user) {
+        final Maybe<UserProvider> userProvider = identityProviderManager.getUserProvider(user.getSource());
+        return userProvider.isEmpty().flatMap(noProvider -> noProvider ?
+                Single.just(updateUser) :
+                updateWithUserProvider(updateUser, user, userProvider.toSingle())
+        );
+    }
+
+    private Single<UpdateUser> updateWithUserProvider(UpdateUser updateUser, User
+            user, Single<UserProvider> userProvider) {
+        return userProvider.flatMap(provider -> provider.findByUsername(user.getUsername())
+                .switchIfEmpty(Single.error(() -> new UserNotFoundException(user.getUsername())))
+                .flatMap(idpUser -> provider.update(idpUser.getId(), convert(user.getUsername(), updateUser)))
+                .map(idpUser -> {
+                    updateUser.setExternalId(idpUser.getId());
+                    return updateUser;
+                })).onErrorResumeNext(ex -> {
+            if (ex instanceof UserNotFoundException) {
+                // idp user does not exist, return the updated user to save in AM
+                return Single.just(updateUser);
+            }
+            return Single.error(ex);
+        });
+    }
+
+    @Override
+    public Single<User> updateStatus(ReferenceType referenceType, String referenceId, String userId, boolean status, io.
+            gravitee.am.identityprovider.api.User principal) {
+        return getUserService().findById(referenceType, referenceId, userId)
+                .flatMap(user -> {
+                    user.setEnabled(status);
+                    return getUserService().update(user);
+                })
+                .doOnSuccess(user1 -> auditService.report(AuditBuilder.builder(UserAuditBuilder.class).principal(principal).type((status ? EventType.USER_ENABLED : EventType.USER_DISABLED)).user(user1)))
+                .doOnError(throwable -> auditService.report(AuditBuilder.builder(UserAuditBuilder.class).principal(principal).type((status ? EventType.USER_ENABLED : EventType.USER_DISABLED)).reference(new Reference(referenceType, referenceId)).throwable(throwable)));
+    }
+
+    protected io.gravitee.am.identityprovider.api.User convert(AbstractNewUser newUser) {
+        DefaultUser user = new DefaultUser(newUser.getUsername());
+        user.setCredentials(newUser.getPassword());
+
+        Map<String, Object> additionalInformation = new HashMap<>();
+        if (newUser.getFirstName() != null) {
+            user.setFirstName(newUser.getFirstName());
+            additionalInformation.put(StandardClaims.GIVEN_NAME, newUser.getFirstName());
+        }
+        if (newUser.getLastName() != null) {
+            user.setLastName(newUser.getLastName());
+            additionalInformation.put(StandardClaims.FAMILY_NAME, newUser.getLastName());
+        }
+        if (newUser.getEmail() != null) {
+            user.setEmail(newUser.getEmail());
+            additionalInformation.put(StandardClaims.EMAIL, newUser.getEmail());
+        }
+        if (newUser.getAdditionalInformation() != null) {
+            newUser.getAdditionalInformation().forEach(additionalInformation::putIfAbsent);
+        }
+        user.setAdditionalInformation(additionalInformation);
+
+        return user;
+    }
+
+    protected User transform(NewOrganizationUser newUser, ReferenceType referenceType, String referenceId) {
+        User user = new User();
+        user.setId(RandomString.generate());
+        user.setExternalId(newUser.getExternalId());
+        user.setReferenceId(referenceId);
+        user.setReferenceType(referenceType);
+        user.setClient(newUser.getClient());
+        user.setEnabled(newUser.isEnabled());
+        user.setUsername(newUser.getUsername());
+        user.setFirstName(newUser.getFirstName());
+        user.setLastName(newUser.getLastName());
+        user.setEmail(newUser.getEmail());
+        user.setSource(newUser.getSource());
+        user.setInternal(newUser.isInternal());
+        user.setPreRegistration(newUser.isPreRegistration());
+        user.setRegistrationCompleted(newUser.isRegistrationCompleted());
+        user.setPreferredLanguage(newUser.getPreferredLanguage());
+        user.setAdditionalInformation(newUser.getAdditionalInformation());
+        user.setForceResetPassword(newUser.getForceResetPassword());
+        user.setCreatedAt(new Date());
+        user.setUpdatedAt(user.getCreatedAt());
+        if (newUser.getLastPasswordReset() != null) {
+            if (newUser.getLastPasswordReset().toInstant().isAfter(Instant.now())) {
+                throw new UserInvalidException("lastPasswordReset cannot be in the future");
+            }
+            user.setLastPasswordReset(newUser.getLastPasswordReset());
+        } else if(!newUser.isPreRegistration()){
+            user.setLastPasswordReset(user.getCreatedAt());
+        }
+        user.setServiceAccount(newUser.isServiceAccount());
+        return user;
+    }
+
+    protected void updateInfos(User user, NewOrganizationUser newUser) {
+        user.setFirstName(newUser.getFirstName());
+        user.setLastName(newUser.getLastName());
+        user.setEmail(newUser.getEmail());
+        user.setAdditionalInformation(newUser.getAdditionalInformation());
+    }
+
+    protected io.gravitee.am.identityprovider.api.User convert(String username, UpdateUser updateUser) {
+        // update additional information
+        DefaultUser user = new DefaultUser(username);
+        Map<String, Object> additionalInformation = new HashMap<>();
+        if (updateUser.getFirstName() != null) {
+            user.setFirstName(updateUser.getFirstName());
+            additionalInformation.put(StandardClaims.GIVEN_NAME, updateUser.getFirstName());
+        }
+        if (updateUser.getLastName() != null) {
+            user.setLastName(updateUser.getLastName());
+            additionalInformation.put(StandardClaims.FAMILY_NAME, updateUser.getLastName());
+        }
+        if (updateUser.getEmail() != null) {
+            user.setEmail(updateUser.getEmail());
+            additionalInformation.put(StandardClaims.EMAIL, updateUser.getEmail());
+        }
+        if (updateUser.getAdditionalInformation() != null) {
+            updateUser.getAdditionalInformation().forEach(additionalInformation::putIfAbsent);
+        }
+        user.setAdditionalInformation(additionalInformation);
+        if (updateUser.getForceResetPassword() != null) {
+            user.setForceResetPassword(updateUser.getForceResetPassword());
+        }
+        return user;
+    }
+
+    protected io.gravitee.am.identityprovider.api.User convert(User user) {
+        DefaultUser idpUser = new DefaultUser(user.getUsername());
+        idpUser.setCredentials(user.getPassword());
+
+        Map<String, Object> additionalInformation = new HashMap<>();
+        if (user.getFirstName() != null) {
+            idpUser.setFirstName(user.getFirstName());
+            additionalInformation.put(StandardClaims.GIVEN_NAME, user.getFirstName());
+        }
+        if (user.getLastName() != null) {
+            idpUser.setLastName(user.getLastName());
+            additionalInformation.put(StandardClaims.FAMILY_NAME, user.getLastName());
+        }
+        if (user.getEmail() != null) {
+            idpUser.setEmail(user.getEmail());
+            additionalInformation.put(StandardClaims.EMAIL, user.getEmail());
+        }
+        if (user.getAdditionalInformation() != null) {
+            user.getAdditionalInformation().forEach(additionalInformation::putIfAbsent);
+        }
+        idpUser.setAdditionalInformation(additionalInformation);
+        idpUser.setForceResetPassword(Boolean.FALSE);
+        return idpUser;
     }
 }
