+++ gravitee-am-management-api/gravitee-am-management-api-service/src/test/java/io/gravitee/am/management/service/OrganizationUserServiceTest.java
@@ -589,7 +589,4 @@ public class OrganizationUserServiceTest {
                         updatedUser.getPassword() == null
         );
     }
-
-
-
 }
